refactor: standardize service import paths and apply minor formatting adjustments across service files
