+++ src/services/user/user.service.ts
@@ -1,11 +1,11 @@
-import httpClient from '../lib/httpClient';
+import httpClient from '../../lib/httpClient'
 
 export async function getUser() {
-  const response = await httpClient.get('/user');
-  return response.data;
+  const response = await httpClient.get('/user')
+  return response.data
 }
 
 export async function createDependentProfile(profileData) {
-  const response = await httpClient.post('/dependent-profile', profileData);
-  return response.data;
+  const response = await httpClient.post('/dependent-profile', profileData)
+  return response.data
 }
